requirements.txt
numpy>=1.26
pandas>=2.2
pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
rapidfuzz>=3.9
pdf2image>=1.17
pytesseract>=0.3.10
easyocr>=1.7.1
openai>=1.45.0
transformers>=4.44
torch>=2.4
